requirement.txt
amqp==2.5.1
ansible==2.8.6
astroid==2.3.2
billiard==3.6.1.0
celery==4.3.0
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.1.5
django-cors-headers==3.1.1
django-crispy-forms==1.8.0
django-debug-toolbar==2.0
django-filter==2.2.0
django-formtools==2.1
django-guardian==2.1.0
django-import-export==1.2.0
django-reversion==3.0.4
djangorestframework==3.10.3
et-xmlfile==1.0.1
future==0.18.1
httplib2==0.14.0
idna==2.8
importlib-metadata==0.23
isort==4.3.21
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10.3
kombu==4.6.5
lazy-object-proxy==1.4.2
Markdown==3.1.1
MarkupPy==1.14
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
odfpy==1.4.0
openpyxl==3.0.0
pexpect==4.7.0
Pillow==6.2.1
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.4.2
pylint==2.4.3
pytz==2019.3
PyYAML==5.1.2
rabbitmq==0.2.0
redis==3.3.11
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
tablib==0.14.0
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.6
vine==1.3.0
wrapt==1.11.2
xlrd==1.2.0
XlsxWriter==1.2.2
xlwt==1.3.0
zipp==0.6.0
